pyproject.toml
[tool.ruff]
# Exclude a variety of commonly ignored directories.
extend-exclude = [
    ".direnv",
    ".git",
    ".ruff_cache",
    "node_modules",
    "migrations",
]
line-length = 120
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
fixable = [
    "ALL",
]
unfixable = []
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "amplifier.settings"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--reuse-db",
    "-v",
]

[tool.uv.sources]
# allow installing django-celery-beat 2.8.1 with django 6.0
# until https://github.com/celery/django-celery-beat/issues/977 is resolved
django-celery-beat = { git = "https://github.com/celery/django-celery-beat", rev = "bd429063cbb227c00905e7693c38895e44ff8e47" }

[tool.setuptools]
py-modules = []

[project]
name = "amplifier"
version = "0.1.0"
description = "AMPER is a top-notch, next-gen SFA/FFM/e-commerce solution."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "Django>=6.0",
    "boto3",
    "celery",
    "celery-progress",
    "celery[redis]",
    "channels",
    "channels_redis",
    "cryptography",
    "django-allauth[socialaccount,mfa,headless]>=65.13.0",
    "django-autoslug",
    "django-anymail[mailgun]", # email services
    "django-browser-reload",
    "django-celery-beat",
    "django-cors-headers",
    "django-environ",
    "django-health-check",
    "django-hijack",
    "django-htmx",
    "django-storages",
    "django-ckeditor-5",
    "django-unfold>=0.76.0",
    "django-vite>=3.1.0",
    "django-watchfiles",
    "djangorestframework",
    "djangorestframework-api-key",
    "drf-spectacular",
    "pillow>=12.1.0",
    "psycopg2-binary",
    "redis",
    "tinycss2>=1.5.1",
    "django-import-export>=4.4.0",
    "django-simple-history",
    "django-colorfield>=0.14.0",
]

[dependency-groups]
dev = [
    "channels[daphne]",
    "django-debug-toolbar",
    "pre-commit",
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
    "ruff>=0.13.1",
]
prod = [
    "gunicorn",
    "uvicorn[standard]",
]
